chore: WIP defaultValue async arrays
