+++ src/admin/components/forms/Form/buildStateFromSchema/buildStateFromSchema.spec.js
@@ -1,4 +1,4 @@
-import buildStateFromSchema from './buildStateFromSchema';
+import buildStateFromSchema from './index';
 
 describe('Form - buildStateFromSchema', () => {
   const defaultValue = 'Default';
